+++ Plumber.build.ps1
@@ -4,6 +4,12 @@
         PSScriptAnalyzer = @{
             Exclude = @('Tests/Assets/TaskHelp/InvalidPowerShell.ps1')
         }
+        PathSeparator    = @{
+            Exclude = @(
+                'Tests/Assets/TaskHelp/InvalidPowerShell.ps1'
+                'Tests/Unit/Private/Get-PlumberPathSeparator.Tests.ps1'
+            )
+        }
         Local            = @(
             'LocalTasks/ValidateTaskHelp.ps1'
         )