+++ Tasks/ModuleConventions/Naming.ps1
@@ -36,7 +36,7 @@
     ```
 #>
 Add-BuildTask -Name Naming -Jobs SetVariables, {
-    $psm1Name = Get-Item "$BuildRoot\$script:moduleName.psm1"
+    $psm1Name = Get-Item (Join-Path $BuildRoot "$script:moduleName.psm1")
     if (-not ($script:psd1.RootModule -ceq $psm1Name.Name)) {
         Write-Error "$($psm1Name.Name) case does not match RootModule"
     }